refactor: extract unit save into a hook
